backend/requirements/development.txt
-r _base.txt
brunette==0.2.2
django-coverage-plugin==2.0.2
django-extensions==3.1.5
django-stubs==1.9.0
djangorestframework-stubs==1.4.0
mypy==0.910
pre-commit==2.17.0
pylint==2.12.2
pylint-celery==0.3
pylint-django==2.4.4
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-factoryboy==2.1.0
